Fix 32-bit Windows warning with explicit cast

+++ clients/drcachesim/tools/schedule_stats.h
@@ -230,7 +230,7 @@ public:
         {
             // Ensure {workload_id=X, tid=Y} doesn't always hash the same as
             // {workload_id=Y, tid=X} by avoiding a simple symmetric wid^tid.
-            return std::hash<size_t>()(wt.workload_id ^ wt.tid) ^
+            return std::hash<size_t>()(static_cast<size_t>(wt.workload_id ^ wt.tid)) ^
                 std::hash<memref_tid_t>()(wt.tid);
         }
     };
